+++ server/src/resources/tapestries.ts
@@ -11,8 +11,7 @@ import {
   TapestryDto,
   TapestryUpdateDto,
 } from 'tapestry-shared/src/data-transfer/resources/dtos/tapestry.js'
-import { queue } from '../tasks/index.js'
-import { config } from '../config.js'
+import { scheduleTapestryThumbnailGeneration } from '../tasks/utils.js'
 import { BadRequestError, ForbiddenError } from '../errors/index.js'
 import { ListParamsOutputDto } from 'tapestry-shared/src/data-transfer/resources/dtos/common.js'
 import { createItems } from './items.js'
@@ -117,23 +116,6 @@ async function orderByInteraction(
   }
 
   return tapestries
-}
-
-export async function scheduleTapestryThumbnailGeneration(
-  tapestryId: string,
-  { skipDelay = false } = {},
-) {
-  await queue.remove(tapestryId)
-  await queue.add(
-    'generate-tapestry-thumbnails',
-    { tapestryId },
-    {
-      jobId: tapestryId,
-      delay: skipDelay ? 0 : config.worker.tapestryThumbnailGenerationDelay,
-      removeOnComplete: true,
-      removeOnFail: true,
-    },
-  )
 }
 
 function fecthTapestryByPathSegment(
