+++ server/src/tasks/process-item-thumbnail.ts
@@ -1,11 +1,10 @@
 import { s3Service, tapestryKey } from '../services/s3-service.js'
 import { ensureTransaction, prisma } from '../db.js'
 import { ImageAssetRendition, Item, Prisma } from '@prisma/client'
-import { scheduleTapestryThumbnailGeneration } from '../resources/tapestries.js'
 import { ThumbnailRenditionOutput } from './thumbnail-generators/index.js'
 import { generateItemThumbnailRenditionName } from 'tapestry-shared/src/utils.js'
 import { parseDBItemSource } from '../transformers/item.js'
-import { downloadImageToArrayBuffer } from './utils.js'
+import { downloadImageToArrayBuffer, scheduleTapestryThumbnailGeneration } from './utils.js'
 import { generateThumbnail } from './thumbnail-generators/image.js'
 import sharp from 'sharp'
